Refactor: Single Activity Architecture로 구조변경을 위해 리팩토링
